Move Agency Mapper Classes
- ImageMapper
- PostMapper
- SectionMapper
- TagMapper
